+++ rust/src/cli.rs
@@ -1,11 +1,12 @@
-use crate::config;
-use crate::decode;
+use crate::convert;
 use crate::open;
 use crate::plugins;
 use crate::request;
 use crate::serve;
 use crate::upgrade;
 use crate::validate;
+use crate::{config, plugins::read_plugins};
+use anyhow::Result;
 use structopt::StructOpt;
 use strum::{Display, EnumVariantNames};
 
@@ -26,7 +27,7 @@ pub struct Args {
 )]
 pub enum Command {
     Open(open::cli::Args),
-    Decode(decode::cli::Args),
+    Convert(convert::cli::Args),
     Validate(validate::cli::Args),
     Serve(serve::cli::Args),
     Request(request::cli::Args),
@@ -35,18 +36,24 @@ pub enum Command {
     Config(config::cli::Args),
 }
 
-// TODO Return a result?
-pub async fn cli(args: Vec<String>) -> i32 {
-    // TODO: Do not call updrade if command itself is upgrade
-    let upgrade_thread = upgrade::upgrade_auto();
-
+pub async fn cli(args: Vec<String>) -> Result<i32> {
     // Parse args into a command
     let Args { command } = Args::from_iter(args);
+
+    // If not explicitly upgrading then run an upgrade check in the background
+    let upgrade_thread = if let Command::Upgrade(_) = command {
+        None
+    } else {
+        Some(upgrade::upgrade_auto())
+    };
+
+    // Load plugins
+    read_plugins()?;
 
     // Run the command
     let result = match command {
         Command::Open(command) => open::cli::open(command).await,
-        Command::Decode(command) => decode::cli::decode(command),
+        Command::Convert(command) => convert::cli::convert(command),
         Command::Validate(command) => validate::cli::validate(command),
         Command::Serve(command) => serve::cli::serve(command).await,
         Command::Request(command) => request::cli::request(command).await,
@@ -56,17 +63,19 @@ pub async fn cli(args: Vec<String>) -> i32 {
     };
 
     // Join the upgrade thread and log any errors
-    if let Err(_error) = upgrade_thread.join() {
-        // TODO: Log error
+    if let Some(upgrade_thread) = upgrade_thread {
+        if let Err(_error) = upgrade_thread.join() {
+            // TODO: Log error
+        }
     }
 
     match result {
-        Ok(_) => exitcode::OK,
+        Ok(_) => Ok(exitcode::OK),
         Err(error) => {
             // Write the error to the terminal
             // TODO Send this to a logger
             eprintln!("{}", error);
-            exitcode::SOFTWARE
+            Ok(exitcode::SOFTWARE)
         }
     }
 }